Golang Server For websockets proxy and serving static assets (#4)
## GoLang Server

This spins up a GoLang server that

1. Handles websockets from the client
2. Invokes Runme to handle requests from the client

This eliminates the typescript backend that was previously used in #2.
This way we don't need to have 2 servers to serve the backend.

## Replace socket.io with plain websockets

I ended up removing socket.io in the client library and going back to
NativeWebsockets. I did this because it looks like the GoLang server is
only using 1.4 version of the socket.io "protocol" but the client is on
4.x and I saw that you shouldn't expect compatibility.

I was reading this about the benefits of socket.io
https://socket.io/docs/v4/#is-socketio-still-needed-today. IIRC I think
you were having trouble with reconnections. I'm hoping maybe using
GoLang we