Introduce optional OIDC authentication (#39)
- [x] Use OIDC token to auth sessions
- [x] No OIDC config disables auth (this is handy for dev when
frontend's running via `vite`)
- [x] Only allow tokens that match a list of hosted domains `hd` claim
- [x] Download provider keypair at startup to verify signatures offline,
skip refresh
- [x] Leverage `httpOnly` cookie for all requests/WS from the browser
for authorization
- [x] Use JWT parser claims
- [x] Init random state for added security
- [ ] The UX needs improvement
- [x] Add some tests

Unsolved: OIDC token JWTs are only valid for `1h`; that's too short to
force re-login every time.

The simplest solution would be to generate a keypair at startup and
issue session tokens with a lifetime of 12-24 hours, which also
invalidate on restart. Another option is to write the keypair into the
user config dir and reuse it to honor the whole token lifetime.

---------

Signed-off-by: Sebastian (Tiedtke) Huckleberry <sebastiantiedtke@gmail.com>