Generic OIDC to support, e.g. Azure's Entra ID (#41)
This will allow any generic OIDC, not just Google, to be configured.
Successfully tested it with Azure Entra ID (video below).

The F/E needs to be wired up for better error/feedback, etc. Also, when
API calls are rejected due to lack of auth, we should forward them to a
login page and not silently ignore them.

Also added some tests because OIDC can be finicky.


https://github.com/user-attachments/assets/f3fce972-a760-49b0-9a14-0d52b8bf0c9a

---------

Signed-off-by: Sebastian (Tiedtke) Huckleberry <sebastiantiedtke@gmail.com>