AuthZ with OIDC - IAM Policies (#43)
This PR improves the AuthZ logic added in the original OIDC
implementation.

The original PR added the ability to Authorize based on users being in a
given domain. We need finer grained authorization at the individual
level. For example, a runner (e.g. on localhost or devbox) might be
restricted to a single user.

This PR introduces IAM Policies, Roles, and Binding modeled on how GCP
handles IAM. We can then define an IAM policy which grants certain roles
to individual users. Currently two roles are defined runner.user and
agent.user.
Currently members can only be specified as individuals. In the future we
could extend it to all "all users in a domain", "groups" or similar
semantics based on requirements.

This PR also starts storing the OIDC token in the context and then
passing that along to a separate Auth middle ware that handles AuthZ.
This allows us to have method specific authorization handling.