Fix Function Calling and Linking to Previous Responses (#84)
Previous response_id is no longer getting set and neither was the
call_id being passed along. As a result, the AI

* Wasn't taking into account previous responses
* The tool call outputs

Both problems were detectable via the manual agent_test but this wasn't
being run in CI and therefore regularly.

It looks like the cause of the problem is that the Responses API changed
and response_id and call_id are now only being set on a subset of
events. We therefore need to modify the code to properly handle this.

It looks like response_id gets set on the initial created_event

https://platform.openai.com/docs/api-reference/responses-streaming/response/created
Previously we were assuming it was set on all events but that no longer
appears to be the case.

To fix this we set the response_id the first time we get an event that
has non empty response_id in the block. We then save it to attach to all
subsequent responses returned by the agent.

For the call_id it looks like it is set on the output_item added event
and item_done events but not the functional_call_argument_delta events.
The API spec isn't super clear here
If you look at the doc about streaming and how to accumulate events you
can see that call id is only on these events and not the deltas.

https://platform.openai.com/docs/guides/function-calling?api-mode=responses#streaming

But it doesn't show up in the API spec

https://platform.openai.com/docs/api-reference/responses-streaming/response/output_item/added

My hypothesis is that this is related to the upgrade in #73 of the
openai go client.

## GHA

Start running the `agent_test` in GHA to catch further regressions.

Related to #76