Support Credential Refreshing (#103)
We want to support OAuth as part of our login flow.

As part of logging in we will store the OAuth token returned by the OIDC
flow. This will be stored in a separate cookie on the client. This will
allow the client to

1. Obtain a refresh token
2. Determine when the token is expiring
3. Automatically refresh the access token and id token using the refresh
token

We store the full OAuth token as a separate cookie containing the
serialized JSON of a proto representation of the cookie. This allows us
to have a consistent, well defined schema across our application.

We use a separate cookie because this allows our existing code path to
remain unchanged. Notably we don't want to set more than the idToken to
the runner.

# List of chaings

* Update OIDC flow to support PKCE
* Allow more than one vector store 

# Alternatives

There are a lot of different choices we could make here e.g.

* Using a single cookie to support both the access token and the id
token
* Passing the  access token as a header

For #1 keeping the ID token separate allows existing code paths to
remain unchanged; in particular nothing needs to change w.r.t. how the
ID token gets passed along to the Agent or Runner for logging in
purposes.

I was on the fence about whether to optionally include the access token
in a header or the response body of the request.

I think the advantage of putting it in the request body is that its an
explicit part of the schema. Downside is we'd have to add it to every
request that needs it but right now we only have one so seems like a
non-issue.